- (snmp_api.c): trap types were int's not longs (failed under 64bit arch).
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@1479 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/snmp_api.c
@@ -183,8 +183,8 @@ struct internal_snmp_pdu {
     oid     *enterprise;/* System OID */
     int     enterprise_length;
     snmp_ipaddr  agent_addr; /* address of object generating trap */
-    int     trap_type;  /* trap type */
-    int     specific_type;  /* specific type */
+    long     trap_type;  /* trap type */
+    long     specific_type;  /* specific type */
     u_long  time;       /* Uptime */
 
     struct variable_list *variables;
